requirements.txt
falcon==1.1.0
colander==1.3.2
google-cloud-datastore==0.22.1
google-api-python-client==1.6.1
pytest==3.0.6

